requirements.txt
langchain==0.1.13
streamlit==1.32.2
tiktoken==0.5.2
python-dotenv==1.0.1
huggingface_hub
langchain_community